Eliminate 120 unnecessary type: ignore suppressions and enforce type safety at boundaries
Systematic overhaul of the 611 type: ignore comments across the codebase.
Stripped all 526 defensive unused-ignore blanket tags that were hiding
redundancies, then used mypy with full library stubs (numpy 2.4, torch 2.10,
scipy 1.17) to identify and remove genuinely unnecessary suppressions and
replace others with proper type-safe code.

Changes by category:
- Removed 73 unused suppressions confirmed unnecessary by mypy
- Fixed 67 no-any-return with explicit float()/str()/bool()/np.asarray()
- Added cast() wrappers for 16 lazy import getters in engine.py
- Added cast() for 3 decorator returns in utils/resilience.py
- Added assert-not-None guards for streaming, routing, emergent detectors
- Replaced float(object) with float(str(x)) for 5 dict-access patterns
- Proper type alias annotation for AvaDominanceEquation

Result: 611 → 491 type: ignore comments (20% reduction). Every remaining
suppression has a specific mypy error code documenting exactly what it
suppresses. Zero unused-ignore blankets remain.

+++ src/omni_mercury_engine/core/three_r/learnable_fusion.py
@@ -45,9 +45,9 @@ try:
     TORCH_AVAILABLE = True
 except ImportError:
     TORCH_AVAILABLE = False
-    torch = None  # type: ignore[assignment, unused-ignore]
-    nn = None  # type: ignore[assignment, unused-ignore]
-    F = None  # type: ignore[assignment, unused-ignore]
+    torch = None  # type: ignore[assignment]
+    nn = None  # type: ignore[assignment]
+    F = None  # type: ignore[assignment]
 
 
 from omni_mercury_engine.core.three_r.types import (
@@ -154,7 +154,7 @@ if TORCH_AVAILABLE:
             raw_gate = self.gate_network(scores)
             scaled_gate = self.min_gate + raw_gate * (self.max_gate - self.min_gate)
 
-            return scaled_gate  # type: ignore[no-any-return, unused-ignore]
+            return scaled_gate  # type: ignore[no-any-return]
 
     class MultiScaleRecursion(nn.Module):
         """
@@ -567,8 +567,8 @@ class Learnable3REngine:
                 weight_decay=self.config.weight_decay,
             )
         else:
-            self.model = None  # type: ignore[assignment, unused-ignore]
-            self.optimizer = None  # type: ignore[assignment, unused-ignore]
+            self.model = None  # type: ignore[assignment]
+            self.optimizer = None  # type: ignore[assignment]
 
         self.training_history: list[float] = []
 
@@ -644,7 +644,7 @@ class Learnable3REngine:
 
         loss = F.mse_loss(prediction, target_tensor)
 
-        loss.backward()  # type: ignore[no-untyped-call, unused-ignore]
+        loss.backward()  # type: ignore[no-untyped-call]
         self.optimizer.step()
 
         loss_value = float(loss.item())
@@ -662,7 +662,7 @@ class Learnable3REngine:
             mid = len(x) // 2
             left = recursion_score(x[:mid], depth - 1)
             right = recursion_score(x[mid:], depth - 1)
-            return 0.5 * (left + right) + 0.5 * np.std(x) / (np.mean(np.abs(x)) + 1e-8)  # type: ignore[no-any-return]
+            return float(0.5 * (left + right) + 0.5 * np.std(x) / (np.mean(np.abs(x)) + 1e-8))
 
         fft = np.fft.fft(arr)
         magnitudes = np.abs(fft)
